Chappel Corrections
I had to change a lot in script.
The verification in the secretDoor works.

Still needs to add the undo bottun (i alread have the script from the kitchen to use as base)
Need to do the candleholder and the music paper + book to know the places.

Statues of angels dont work. I had to use a different statue to test (lion) and it works well the chalice or torchs.

The verification is done for now with only one correct.